+++ src/main/client/client/GUI/Helpers/ErrorHandlers.java
@@ -1,9 +1,16 @@
 package client.GUI.Helpers;
 
+import client.GUI.CPSClientGUI;
 import javafx.application.Platform;
 import javafx.scene.control.Alert;
+import javafx.scene.control.Label;
+import javafx.scene.control.TextArea;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.Priority;
 
 import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 
 /**
  * Custom error handlers.
@@ -21,15 +28,38 @@ public class ErrorHandlers {
     {
         Alert guiError = new Alert(Alert.AlertType.ERROR);
         guiError.setHeaderText("An error occurred displaying the GUI!");
-        guiError.setContentText("The following IO error occurred while trying to display the interface:\n"
-                + ex.getMessage());
+        guiError.setContentText("An error occurred displaying the GUI!");
+        StringWriter sw = new StringWriter();
+        PrintWriter pw = new PrintWriter(sw);
+        ex.printStackTrace(pw);
+
+        TextArea textArea = new TextArea(sw.toString());
+        textArea.setEditable(false);
+        textArea.setWrapText(true);
+
+        textArea.setMaxWidth(Double.MAX_VALUE);
+        textArea.setMaxHeight(Double.MAX_VALUE);
+        GridPane.setVgrow(textArea, Priority.ALWAYS);
+        GridPane.setHgrow(textArea, Priority.ALWAYS);
+
+        Label label = new Label("The exception stacktrace was:");
+        GridPane expContent = new GridPane();
+        expContent.setMaxWidth(Double.MAX_VALUE);
+        expContent.add(label, 0, 0);
+        expContent.add(textArea, 0, 1);
+
+        guiError.getDialogPane().setExpandableContent(expContent);
+
         Platform.runLater(new Runnable() {
             @Override
             public void run() {
                 guiError.showAndWait();
             }
         });
-        ex.printStackTrace();
+        if (CPSClientGUI.IS_DEBUG)
+        {
+            ex.printStackTrace();
+        }
         if (exit)
         {
             System.exit(-1);